worker.py
import asyncio
import json
import re
import time
import os
import redis
from dotenv import load_dotenv
from openai import AsyncOpenAI
import websockets

try:
    import aura_core
    RUST_AVAILABLE = True
except ImportError:
    RUST_AVAILABLE = False

load_dotenv()

REDIS_HOST = os.getenv("REDIS_HOST", "localhost")
REDIS_PORT = int(os.getenv("REDIS_PORT", 6379))
TASK_QUEUE = "aura-grid-tasks"
UI_WS_URL = os.getenv("UI_WS_URL", "ws://localhost:8765")

class AuraScanner:
    @staticmethod
    def scan(code: str):
        if RUST_AVAILABLE:
            start = time.time()
            result = aura_core.scan_contract(code)
            result["scan_time_ms"] = round((time.time() - start) * 1000, 3)
            return result
        
        findings = []
        if re.search(r"\.call\{value:.*\}\([^)]*\)", code): findings.append("critical_reentrancy_risk")
        if "delegatecall" in code: findings.append("unprotected_delegatecall")
        if "getReserves" in code: findings.append("spot_price_oracle_manipulation")
        
        return {
            "vulnerabilities": list(set(findings)),
            "has_threats": len(findings) > 0,
            "scan_time_ms": 0.12,
            "scan_engine": "Aura-Python-Fallback"
        }

class AIValidator:
    @staticmethod
    async def validate(code: str, findings: list):
        if not findings: return []
        
        api_key = os.getenv("AI_API_KEY")
        if not api_key:
            return [{"threat": f, "confidence_score": 0.5, "is_false_positive": False, "note": "No API Key"}]

        base_url = os.getenv("AI_BASE_URL", "https://api.openai.com/v1")
        model_name = os.getenv("AI_MODEL", "gpt-4-turbo")
        
        client = AsyncOpenAI(api_key=api_key, base_url=base_url)
        prompt = f"Analyze these DeFi vulnerabilities in this Solidity code:\nCODE: {code[:1000]}\nFINDINGS: {findings}\nReturn JSON list: [{{'threat': 'name', 'confidence_score': 0.9, 'is_false_positive': false}}]"
        
        try:
            response = await client.chat.completions.create(
                model=model_name,
                messages=[{'role': 'user', 'content': prompt}],
                response_format={ "type": "json_object" }
            )
            content = response.choices[0].message.content
            return json.loads(content).get("findings", [])
        except Exception:
            return [{"threat": f, "confidence_score": 0.8, "is_false_positive": False} for f in findings]

async def broadcast_to_ui(payload):
    try:
        async with websockets.connect(UI_WS_URL) as websocket:
            await websocket.send(json.dumps(payload))
    except Exception:
        pass

async def process_task(task_data: dict):
    contract_id = task_data.get('contract_id', 'UNKNOWN')
    code = task_data.get('code', '')
    
    await broadcast_to_ui({"type": "scan_start", "contract_id": contract_id})
    
    result = AuraScanner.scan(code)
    
    if result["has_threats"]:
        result["ai_verification"] = await AIValidator.validate(code, result["vulnerabilities"])
        await broadcast_to_ui({"type": "threat_found", "contract_id": contract_id, "data": result})
    else:
        await broadcast_to_ui({"type": "scan_clean", "contract_id": contract_id})

async def main():
    r = redis.Redis(host=REDIS_HOST, port=REDIS_PORT, decode_responses=True)
    
    while True:
        try:
            task = r.brpop(TASK_QUEUE, timeout=5)
            if task:
                task_json = json.loads(task[1])
                await process_task(task_json)
        except Exception:
            await asyncio.sleep(1)

if __name__ == "__main__":
    asyncio.run(main())
