requirements.txt
# Aura-Grid Python Worker Dependencies
redis>=5.0.1
asyncio>=3.4.3
aiohttp>=3.9.1
openai>=1.12.0
python-dotenv>=1.0.0
websockets>=11.0.3
maturin>=1.0.0
